modal mostly works again
